nat_requirements.txt
aeh.earlytest===1.3.0-3-g1e2bbec
alembic==0.8.0
avrc.redcross===1.2.2-4-g8d7c628
Babel==2.0
beautifulsoup4==4.4.0
Chameleon==2.22
coverage==3.7.1
python-dateutil==2.6.0
ddt==1.0.0
funcsigs==0.4
httmock==1.2.3
ldappool==1.0
linecache2==1.0.0
lingua==3.11
Mako==1.0.1
MarkupSafe==0.23
mock==1.3.0
nose==1.3.7
PasteDeploy==1.5.2
pbr==1.5.0
polib==1.0.7
Pycap==1.0.2
Pygments==2.0.2
pyopenssl==0.14
pyramid==1.5.7
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.1
pyramid-ldap==0.1
pyramid-mako==1.0.2
pyramid-rewrite==0.2
pyramid-tm==0.12
python-editor==0.3
python-gnupg==0.3.7
python-ldap==2.4.20
pytz==2015.4
repoze.lru==0.6
requests==2.7.0
semantic-version==2.3.1
six==1.9.0
SQLAlchemy==1.0.8
tabulate==0.7.5
traceback2==1.4.0
transaction==1.4.4
translationstring==1.3
TurboMail==3.0.3
unittest2==1.1.0
venusian==1.0
waitress==0.8.9
WebOb==1.5.0a1
WebTest==2.0.18
Werkzeug==0.10.4
wheel==0.24.0
xlrd==0.9.4
xlutils==1.7.1
xlwt==1.0.0
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.6
